Bug fixes, improvements and added features
Fixed some bugs regarding clearing the stack. The timing wasn't right. We need to make sure that we have generated all assembly for the current block and then clear the stack. Otherwqise we have no way of returning a local variable on the stack.

Also added some functions to separate clearing the stack and emiting the asm code for it.

ANd now we can return integer variables. Like return a; or return b;

These are noit supported yet:
return a +*/- b;
return a + 123;
and so on. I have very limited time to work on this compiler. Almost once every 6 months :D

But if I really want to be serious about it, I have to rewrite a lot of stuff and add a pro[er lexer and make a bytecode representation plus optimizations.